feat(disputes): score_dispute_api submit + admin review (#1012)
User score-dispute API (submit/mine/admin-list/admin-resolve),
record-only. Reuses real data layer + existing auth; self-test passes.
Also fixes McpScoreDispute PK BigInteger->Integer (autoincrements on
sqlite+pg). The module_from_exemplar ladder got ~90% (correct real-model
module) but tripped the self-test on this exact PK-autoincrement nuance;
finished in-session.

+++ migrations/versions/0003_score_disputes.py
@@ -16,7 +16,7 @@ depends_on = None
 def upgrade():
     op.create_table(
         "mcp_score_disputes",
-        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
         sa.Column("server_id", sa.String(length=128), nullable=False, index=True),
         sa.Column("submitted_by", sa.String(length=128), nullable=False, index=True),
         sa.Column("proposed_overall_risk", sa.String(length=16), nullable=False),
